Update drinkmaskin.py
Changed the initial output from LOW to HIGH and the output in respective drink. The signal seems to be inverted, since setting the initial output as LOW made the pumps run constantly, until given the signal LOW (or False)..

I've trimmed the times to make good drinks, and also made two drinks that is merely booze.